backup/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
PyMuPDF>=1.22.0
openai>=1.3.0
python-dateutil>=2.8.2
pytz>=2023.3

# Development dependencies
pytest>=7.4.0
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0
